refactor(racingcar) : 자동차 객체가 스스로 판단하여 이동하도록 기능 변경
