+++ docs/README.md
@@ -7,15 +7,11 @@
 - [ ] 입력 : InputView
   - [ ] 경주할 자동차들의 이름을 쉼표 기준으로 입력 : getCars
   - [ ] 라운드 횟수 입력 : getRounds
-- [x] 게임 진행 : CarManager
-  - [x] 0~9 사이 무작위 값을 생성 : createRandomNumber
-  - [x] 무작위 값이 4 이상일 경우 전진. count만큼 반복 : moveIfSatisfy
 - [x] 자동차 객체 : Car
   - [x] 인스턴스 변수 : 이름
   - [x] 인스턴스 변수 : 위치
   - [x] 생성자 : 이름은 5자 이하만 가능
-  - [x] 메서드 : 전진
+  - [x] 메서드 : 무작위 값이 4 이상일 경우 전진. count만큼 반복 - moveIfSatisfy
 - [x] 자동차 객체 일급컬렉션 : Cars
   - [x] 인스턴스 변수 : 자동차 리스트
-  - [x] 생성자 : 문자열 입력받아 자동차 리스트 생성
   - [x] 메서드 : 자동차 조회
